+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Add tests for logical vectors: `some()` and `same()`
 - Spell out numeric vectors using `spell_numbers()` and ranges using `spell_range()`
+- Deselect columns filled entirely with `NA` using `deselect_na()`
 - Check function arguments and set up testing
 - Use `stringi::stri_join()` instead of `paste()` and `paste0()`